refactor(entity): split stats.rs into directory module
Promotes the 935-line crates/entity/src/stats.rs (over the 700-line
hard cap in CLAUDE.md) into a stats/ directory with submodules along
its natural type boundaries. Public API unchanged: the new mod.rs
re-exports Stat, StatList, ArchetypeStatValues, and all stat-ID
constants so external callers (cimmeria_entity::stats::HEALTH, etc.)
continue to resolve.

  stats/
  ├── mod.rs           # pub use facade
  ├── stat_ids.rs      # stat ID constants + PUBLIC_STATS allowlist
  ├── stat.rs          # single-stat struct + impl
  ├── stat_list.rs     # collection + serialization helpers
  ├── archetype.rs     # ArchetypeStatValues input struct
  └── tests.rs         # existing unit tests

No behavior change. cargo test -p cimmeria-entity --lib stats: 21
passed.
